In the MySQL optimized query split SELECT / UPDATE
Given the following index on the delayed_jobs table:

CREATE INDEX delayed_job_priority ON delayed_jobs(priority,run_at);

MySQL will ignore this index, and does not support hints when doing an
UPDATE SET ... LIMIT 1; Without an index, the ordering is done via a
filesort, and can be problematic when the delayed_jobs table
becomes large (>10,000 records). This results in an increasingly long
system lock wait time to reserve the job.

By separating the SELECT from the UPDATE, we can take advantage of an
INDEX while fetching the row to lock for a reservation.

More information on SELECT ... FOR UPDATE here:
http://dev.mysql.com/doc/refman/5.7/en/innodb-locking-reads.html

More information on the issue at:
https://mrfrosti.com/2016/06/06/delayed-job-performance-tuning/