Dont use locked_by when looking for job
1. Locked by check is used in cases when the worker dies and then it comes
back to life. Then we would want it to pick up the job he dropped when
dying. In our case we don't name our workers and the default is
host:pid. When the worker dies it is spawn again but its pid will be
different. So there is no (or very little) chance that we will have
worker with the same name. So there is no point for checking for it.
Moreover this check breaks usage of any index we could think of.
It's better for the job to time out and being picked up by another
worker.

2. This commit introduces better indexes for our use case of
Delayed::Job.

3. This commit removes the ordering when updating the job with
locked_at, locker_by.  We don't need  to sort by priority (by run_at)
when we update jobs with locked_by and locked_at as it does not matter
(there will be only 1 result cause we filter by id).

4. When there is only 1 queue in worker's queue list then let's
use equallity. Not sure how good mysql is in optimizing in queries into
equalities if there is only 1 element.